refactor: call `huggingface_download` from module entry
